feature: provide a bold control on memory usage
To avoid constant growth of internal hashtable (holding metadata info,
amongst other things), we can limit its size.  This is currently very
basic, since we still have to implement a purge (LRU? pick a strategy).

We can set this limit with the environment variable:

$ export DROPLETFS_HASHTABLE_MAX_SIZE=10000000

or in the configuration file ($HOME/.dplfsrc):

hashtable_max_size = 10000000

The value is in bytes.  Default value is 10,000,000 (10MB).